+++ src/filter_flags.rs
@@ -1,11 +1,9 @@
 use clap::{Arg, ArgAction};
 
-use crate::filter_flags::FilterFlagType::*;
-
 #[derive(Debug)]
 pub(crate) enum FilterFlagType {
   App,
-  Killed,
+  _Killed,
   Service,
   Started,
   Stopped,
@@ -14,63 +12,48 @@ pub(crate) enum FilterFlagType {
 impl FilterFlagType {
   pub(crate) fn id(&self) -> &'static str {
     match &self {
-      App => "app-flag",
-      Killed => "killed-flag",
-      Service => "service-flag",
-      Started => "started-flag",
-      Stopped => "stopped-flag",
+      Self::App => "app-flag",
+      Self::_Killed => "killed-flag",
+      Self::Service => "service-flag",
+      Self::Started => "started-flag",
+      Self::Stopped => "stopped-flag",
     }
   }
 
-  pub(crate) fn option(&self) -> &'static str {
+  fn option(&self) -> &'static str {
     match &self {
-      App => "app",
-      Killed => "killed",
-      Service => "service",
-      Started => "started",
-      Stopped => "stopped",
+      Self::App => "app",
+      Self::_Killed => "killed",
+      Self::Service => "service",
+      Self::Started => "started",
+      Self::Stopped => "stopped",
     }
   }
 
-  pub(crate) fn shortcut(&self) -> Option<char> {
+  fn shortcut(&self) -> Option<char> {
     match &self {
-      App => None,
-      Killed => None,
-      Service => None,
-      Started => None,
-      Stopped => None,
+      Self::App => None,
+      Self::_Killed => None,
+      Self::Service => None,
+      Self::Started => None,
+      Self::Stopped => None,
     }
   }
 }
 
 pub(crate) fn create_filter_flag(flag_type: &FilterFlagType, subject: &str, long_help: Option<&str>) -> Arg {
   match flag_type {
-    App => app_flag(subject, long_help),
-    Killed => killed_flag(subject, long_help),
-    Service => service_flag(subject, long_help),
-    Started => started_flag(subject, long_help),
-    Stopped => stopped_flag(subject, long_help),
+    FilterFlagType::App => create_clap_filter_flag(FilterFlagType::App, subject, format!("Include apps related to this {}.", subject), long_help),
+    FilterFlagType::_Killed => create_clap_filter_flag(FilterFlagType::_Killed, subject, format!("Include killed {}'s.", subject), long_help),
+    FilterFlagType::Service => create_clap_filter_flag(
+      FilterFlagType::Service,
+      subject,
+      format!("Include services related to this {}.", subject),
+      long_help,
+    ),
+    FilterFlagType::Started => create_clap_filter_flag(FilterFlagType::Started, subject, format!("Include started {}'s.", subject), long_help),
+    FilterFlagType::Stopped => create_clap_filter_flag(FilterFlagType::Stopped, subject, format!("Include stopped {}'s.", subject), long_help),
   }
-}
-
-fn app_flag(subject: &str, long_help: Option<&str>) -> Arg {
-  create_clap_filter_flag(App, subject, format!("Include apps related to this {}.", subject), long_help)
-}
-
-fn killed_flag(subject: &str, long_help: Option<&str>) -> Arg {
-  create_clap_filter_flag(Killed, subject, format!("Include killed {}'s.", subject), long_help)
-}
-
-fn service_flag(subject: &str, long_help: Option<&str>) -> Arg {
-  create_clap_filter_flag(Service, subject, format!("Include services related to this {}.", subject), long_help)
-}
-
-fn started_flag(subject: &str, long_help: Option<&str>) -> Arg {
-  create_clap_filter_flag(Started, subject, format!("Include started {}'s.", subject), long_help)
-}
-
-fn stopped_flag(subject: &str, long_help: Option<&str>) -> Arg {
-  create_clap_filter_flag(Stopped, subject, format!("Include stopped {}'s.", subject), long_help)
 }
 
 fn create_clap_filter_flag(flag_type: FilterFlagType, _subject: &str, help: String, long_help: Option<&str>) -> Arg {