+++ src/modifier_flags.rs
@@ -1,10 +1,7 @@
 use clap::{Arg, ArgAction};
 
-use crate::modifier_flags::ModifierFlagType::*;
-
 #[derive(Debug)]
 pub(crate) enum ModifierFlagType {
-  Json,
   MultiLine,
   Regex,
 }
@@ -12,52 +9,46 @@ pub(crate) enum ModifierFlagType {
 impl ModifierFlagType {
   pub(crate) fn id(&self) -> &'static str {
     match &self {
-      Json => "json-flag",
-      MultiLine => "multi-line-flag",
-      Regex => "regex-flag",
+      Self::MultiLine => "multi-line-flag",
+      Self::Regex => "regex-flag",
     }
   }
 
   pub(crate) fn option(&self) -> &'static str {
     match &self {
-      Json => "json",
-      MultiLine => "multi-line",
-      Regex => "regex",
+      Self::MultiLine => "multi-line",
+      Self::Regex => "regex",
     }
   }
 
   pub(crate) fn shortcut(&self) -> Option<char> {
     match &self {
-      Json => Some('j'),
-      MultiLine => Some('m'),
-      Regex => Some('r'),
+      Self::MultiLine => Some('m'),
+      Self::Regex => Some('r'),
     }
   }
 }
 
 pub(crate) fn create_modifier_flag(flag_type: &ModifierFlagType, subject: &str, long_help: Option<&str>) -> Arg {
   match flag_type {
-    Json => json_flag(subject, long_help),
-    MultiLine => multi_line_flag(subject, long_help),
-    Regex => regex_flag(subject, long_help),
+    ModifierFlagType::MultiLine => create_clap_modifier_flag(
+      ModifierFlagType::MultiLine,
+      subject,
+      format!("Enter the {} as multi-line string.", subject),
+      long_help,
+    ),
+    ModifierFlagType::Regex => create_clap_modifier_flag(
+      ModifierFlagType::Regex,
+      subject,
+      format!(
+        "Interpret the query string as a regular expression instead of an exact matching {} value. \
+         The regular expression syntax is described on \
+         the following web-page: https://docs.rs/regex/latest/regex/#syntax.",
+        subject
+      ),
+      long_help,
+    ),
   }
-}
-
-fn json_flag(subject: &str, long_help: Option<&str>) -> Arg {
-  create_clap_modifier_flag(Json, subject, format!("Show the {} as json.", subject), long_help)
-}
-
-fn multi_line_flag(subject: &str, long_help: Option<&str>) -> Arg {
-  create_clap_modifier_flag(MultiLine, subject, format!("Enter the {} as multi-line string.", subject), long_help)
-}
-
-fn regex_flag(subject: &str, long_help: Option<&str>) -> Arg {
-  create_clap_modifier_flag(
-    Regex,
-    subject,
-    format!("Interpret the query string as a regular expression instead of an exact matching {} value. The regular expression syntax is described on the following web-page: https://docs.rs/regex/latest/regex/#syntax.", subject),
-    long_help,
-  )
 }
 
 fn create_clap_modifier_flag(flag_type: ModifierFlagType, _subject: &str, help: String, long_help: Option<&str>) -> Arg {