+++ src/subjects/env.rs
@@ -77,7 +77,7 @@ impl CommandExecutor for EnvFind {
   async fn execute(&self, target: Option<String>, _: Option<String>, matches: &ArgMatches, context: &Context) -> DshCliResult {
     let query = target.unwrap_or_else(|| unreachable!());
     let query_processor: &dyn QueryProcessor =
-      if matches.get_flag(ModifierFlagType::Regex.id()) { &RegexQueryProcessor::create(query.as_str())? } else { &ExactMatchQueryProcessor::create(query.as_str())? };
+      if matches.get_flag(ModifierFlagType::Regex.id()) { &RegexQueryProcessor::create(&query)? } else { &ExactMatchQueryProcessor::create(&query)? };
     let (include_started, include_stopped) = include_started_stopped(matches);
     context.print_explanation(format!("find environment variables in services that {}", query_processor.describe()));
     let start_instant = Instant::now();